added controllers folder and moved functionality
