--- backend/utility.py
@@ -1,31 +1,0 @@
-# from passlib.context import CryptContext
-# import jwt
-# import os
-# from datetime import datetime, timezone, timedelta
-
-
-# SECRET_KEY = os.getenv("SECRET_KEY")
-# HASHING_ALGORITHM = os.getenv("HASHING_ALGORITHM")
-# ACCESS_TOKEN_EXPIRE_MINUTES = os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES")
-
-# pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-# def hash_password(password: str) -> str:
-#     return pwd_context.hash(password)
-
-
-# def verify_password(plain_password: str, hashed_password: str) -> bool:
-#     return pwd_context.verify(plain_password, hashed_password)
-
-# def create_access_token(data: dict, expires_delta: timedelta = None):
-#     to_encode = data.copy()
-#     if expires_delta:
-#         expire = datetime.now(timezone.utc) + expires_delta
-#     else:
-#         expire = datetime.now(timezone.utc) + timedelta(
-#             minutes=int(ACCESS_TOKEN_EXPIRE_MINUTES)
-#         )
-#     to_encode.update({"exp": expire})
-#     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=HASHING_ALGORITHM)
-#     return encoded_jwt
